deep-scribe/python_backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pyinstaller
chromadb
requests
pytest
pytest-asyncio
